Normalise the variables before storing them in the database

+++ database/101-up-postgres/107-solana/001-init-users.sql
@@ -2,7 +2,7 @@
 
 CREATE TABLE solana_users (
 	address VARCHAR NOT NULL,
-	amount_minted uint256 NOT NULL,
+	amount_minted DOUBLE PRECISION NOT NULL,
 	last_updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
 
 	CONSTRAINT last_minted_non_negative CHECK (amount_minted >= 0)
